+++ src/test/java/nextstep/subway/acceptance/FavoriteAcceptanceTest.java
@@ -3,8 +3,8 @@ package nextstep.subway.acceptance;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.DataLoader;
-import nextstep.member.application.dto.GithubResponses;
+import nextstep.subway.utils.DataLoader;
+import nextstep.subway.stub.GithubResponses;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
